test(wfx-viewer): add some more tests
